No more static utility classes. functions are cool.
